// example/convex/_generated/server.d.ts
type Env = {
  readonly OPENROUTER_API_KEY: string | undefined;
  readonly OPENROUTER_APP_TITLE: string | undefined;
  readonly OPENROUTER_EMBEDDING_DIMENSIONS: string | undefined;
  readonly OPENROUTER_EMBEDDING_MODEL: string | undefined;
  readonly OPENROUTER_HTTP_REFERER: string | undefined;
  readonly OPENROUTER_MODEL: string | undefined;
};
// ...
export declare const env: Env;